requirements.txt
amqp==5.2.0
asgiref==3.7.2
beautifulsoup4==4.12.2
billiard==4.2.0
cachetools==5.3.2
celery==5.3.6
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2023.1
DateTime==5.4
Django==5.0
django-celery-results==2.5.1
django-crispy-forms==2.1
google==3.0.0
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
kombu==5.3.4
prompt-toolkit==3.0.43
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pyserial==3.5
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
requests==2.31.0
rsa==4.9
six==1.16.0
sms==0.4
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
zope.interface==6.1
coverage
django-extensions
django-generic-m2m
django-gm2m
django-querysetsequence
django-nested-admin
django-taggit
django-tagging
mock
selenium
pytest
splinter
pytest-django
pytest-cov
pytest-splinter
tenacity
whitenoise
django-debug-toolbar
sqlparse>=0.4.2 # not directly required, pinned by Snyk to avoid a vulnerability
djhacker
setuptools>=65.5.1 # not directly required, pinned by Snyk to avoid a vulnerability
